requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
crypto==1.4.1
cryptography==2.7
dateutils==0.6.6
Django==2.2.4
django-extensions==2.2.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
mailjet-rest==1.3.3
Naked==0.1.31
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
shellescape==3.4.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
Werkzeug==0.15.5
